make swipe logic more reliable
Before, if you were holding on one side of a swipe at the beginning of
the hit window, that was designated as the "start" of the swipe, which
could make it impossible to hit the swipe if that was on the right
side.

Now, we continuously track the minimum average offset across the life
of the swipe, so that you're able to e.g. hold the right side of a
right swipe, then press the left side with your other hand and swipe.

This also changes how "offset" and "average offset" work in the swipe
algorithm. Before, they were always counted from the left of the note,
so CW swipes would decrease in average offset and CCW swipes would
increase. Now, offset always increases in the direction of the swipe,
so it is counted from the right side of the note for CW swipes. This
simplifies the logic and means that "min average offset" makes way
more sense.