fix bug where one jacket would be wrong
Before, the jacket that is 6 songs to the left of the initially selected song
would always incorrectly show the same jacket as the initially selected song.
This is because `cardHalfCount` is unset when the jackets are first loaded,
because `SongSelectLogic.Awake` is called before `SongSelectCardAnimator.Awake`.
This _only_ affects the very leftmost preview card (the i=0 case in
`LoadAllCards()`) because the loop is async, so in future iterations of the
loop, `SongSelectCardAnimator.Awake` has already run and so now `cardHalfCount`
is actually set properly.

The fix is just to avoid setting `cardHalfCount` during Awake.