don't hit ChainNotes early
Before, ChainNotes could be hit at any time in the window, which usually meant
that they were hit at the very beginning of the window whenever the player was
holding down before the note arrived.

Now, ChainNotes will remember if they've been touched at any point in the early
window, but won't actually be Hit until at least the exact time of the note.
This means that the judgement and hitsound will now play at the right time.