refactor ChartManager and ScoringManager
Most of the logic in ChartManager was moved into ChartLoader, which allows the
decoupling of the chart load logic from the manager that actually is holding the
currently active chart. The new ChartManager is very simple, it just loads the
currently selected chart and stores it in a public field. We don't need any
locks or anything like that because the a new ChartLoader is instantiated every
time we load a chart, avoiding statefulness issues, and the ChartManager just
has a reference to the loaded Chart object, without any additional state which
could cause races.

This allows us to put the new ChartManager only in the gameplay scene, instead
of it being a PersistentSingleton. The information about the currently selected
chart is stored in PersistentStateManager anyway, and this is what the
ChartManager reads from. (Note: once we need to get the chart preview working in
song select, that will probably also require an instance of chart manager. I'm
not worrying about this just yet.)

Also, move the preprocessing that ScoringManager does into the ChartManager -
this allows us to simplify the stupid logic in ScoringManager that tries to
track whenever a chart is loaded.