split up input management responsibility
Before, ScoringManager and InputManager called into each other and replays were
handled by the ScoringManager.

Now, ScoringManager is a pure consumer of inputs. It's the InputManager's job to
provide timestamped inputs to the ScoringManager based on whatever input method
we are using.

The TouchRingManager is created as a PersistentSingleton (needed in the future
for menuing etc) that handles the serial communication with the touch ring
input. The ReplayManager is also created, to manage replay reading and writing.
Both of these implement IInputProvider, which means the InputManager can hook
into them to get new inputs and provide them back to the ScoringManager.

This means that the ScoringManager now can be completely unaware of the current
gameplay clock, the input method, replay handling, etc. It just receives
timestamped inputs and judges them.

As well, keybeams can now hook directly into the InputManager and show keybeams
correctly when playing a replay, without any weird shenanigans.