update scoring engine to support sparse input
This changes the scoring engine so that it will function correctly even when
HandleInput is not called frequently, and only when the input actually changes.
This allows replays to avoid encoding every single frame, and only encode the
frames where input changes - it's assumed that the current input is held until
there is a new frame.

This also introduces a new feature in HandleInput that allows you to call it
with a null TouchState - this is equivalent to saying "we are holding the input
until at least this timeMs", which allows judgement information to update
without providing the next TouchState yet.