simplify input method handling
Previously, input methods would call have a `TouchStateHandler` method which the
`InputManager` would set. Then, whenever there was a new state (that is, every
frame lol), the input method would call the `TouchStateHandler`, which would
then run the code in `InputManager` to handle the input, which would in turn
call the `ScoringManager`, etc...

These levels of indirection were chosen because this model seemed more
well-suited to multi-threaded input, which was not yet implemented but
eventually desired. But this is really confusing and it does not seem like
multi-threaded input will actually be implemented in the near future.

This was drastically simplified so that the keyboard and touch ring input
methods just have a basic method `GetCurrentTouchState` which provides the
TouchState for that frame. Since these input methods are totally synchronous
for now, this is sufficient. Replays take a bit more care to handle well, but
this is now just handled explicitly in the `InputManager`, which is less
confusing.