add artificial latency to InputManager
Create a queue of timed inputs which can be used to artificially introduce
latency into the input system. When latencyMs = 0, this will functional the same
as having no queue at all.

In the future, the queue approach could be used to enable multi-threaded input.
If the queue is threadsafe, then the input thread can write inputs into the
queue with exact timestamps, and the main thread can read the inputs and judge
them exactly. A threadsafe queue can be created without locks by using atomic
operations - see System.Collections.Concurrent.ConcurrentQueue<T>. (We can't use
this directly because of the unique concern around preallocating TouchStates.)