add support for enum settings and make the parameter setting dynamic
Before, GetParameter and SetParameter in PlayerSettings were basically a giant
if statement on the name of the parameter. Now, this is all handled dynamically
using reflection.

Also, allow for parameters to be enums. Some parameters will be changed to use
this in a future commit. This will help make the upcoming settings TOML more
usable.

OffsetMode is changed to use an enum already, since it's easier to just force
any ConditionParameter to be an enum.