finish setting up offset settings
AudioOffset is split into AudioOffsetMs (used in Standard and Advanced mode) and
ClassicOffset (used in classic mode). The split is useful because the units are
different on these two settings - AudioOffsetMs is of course just in
milliseconds, but ClassicOffset is using the arbitrary units from the original
game, from -10.0 to +10.0 in 0.1 increments.

Also change ClassicOffset to use a float instead of the old AudioOffset which
used an int, but was stored as 10x the actual value. Update all the necessary
other code paths to support float settings.

Rename VideoOffset and InputLatency to include the Ms suffix for clarity.

Finally, plumb all these values so that they are actually used by the engine.