fix some rendering quirks with a high visual offset
Previously, notes would disappear as soon as "IsHit" became true. However, IsHit
wasn't very accurate to the colloquial notion of a note being "hit", since a
Miss was also considered hit. Rename IsHit to HitWindowsEvaluated (sorry for the
long name) and introduce a new property HasBeenHit that is closer to what you
would expect. Now, notes disappear when HasBeenHit is true, so missed notes do
not disappear.

Also, make sure that holds continue to have the right state if they are still on
screen after they have ended.

Both of these bugs are primarily noticeable when playing (or not
playing/watching a replay) with a high positive visual offset like +500.