persist data from PersistentStateManager to a TOML file
This generalizes the class that PlayerSettings uses to read/write its data to
TOML, and adds a new toml file called state.toml which tracks the stuff in
PersistentStateManager (except for the LastScoreData, which is really only used
when going from gameplay to the results screen).

The upshot is that the game will now remember what song/diff you had selected in
song selected, as well as which sorting method.